Update for changes in APDB API.
Apdb class is now an abstract interface and its implementation based on
SQL is in ApdbSql class, similar change in ApdbConfig classes.
DiaPipelineConfig is updated to use ApdbSql as a default target, can be
re-targeted to other implementation when it appears. Also there is no
need to override default `isolation_level` in ApdbSql config for sqlite.