For some subtasks only mock out the run() method
With the task mocked out the metadata values calculated
by the task are no longer real values but are themselves
MagicMock objects. MagicMock objects can't be stored in
a TaskMetadata so instead the run methods for the
associator subtasks are explicitly mocked to return
usable data for the metadata.

This works with PropertySet since that seems to accept
the mock value and silently ignores it.