Fix Pandas deprecation warnings.
This commit removes some calls to Dataframe.append that were missed in
3fa22fb7485986c6208eb57453f0ae6d30155669. The calls could not be
detected because they were mocked out.

I don't see a way to prevent such issues short of rewriting the tests to
use real tables in place of mocks, but I don't know how to do so while
following the correct table schema (which is subject to change).