Pass IdFactory to direct forced-phot.
This was a bug, but a very minor one a the IDs in the forced-phot
catalogs aren't really used and probably don't need to exist.  But as
long as they do it's best to make these IDs embed the data IDs like all
other kinds of source IDs (including diffim forced source IDs a few
lines above this change).