Move trailed source filter before diaObject check
If doTrailedResult does not come before diaObjects, then if there are no diaObjects but there are genuine sources that need filtering, they were not being filtered. The filter is moved earlier and then returns an altered diaSources catalog. If the filter is not run, it returns an empty dataframe similar to how matchedDiaSources can also be empty.